➡️ Severity
➡️ High Severity - High Priority bug 
It describes a situation when a major path through the application doesn't work correctly. For example: on an eCommerce website, every customer gets an error message on the booking form and cannot place orders, or the product page throws an Error 500 response.

➡️ High Severity - Low Priority bug 
This happens when the bug causes major problems. But the situation is rare and faces particular conditions. For example, the customers who use deprecated browsers cannot go through the purchasing process. The number of customers with deprecated browsers is low, so this bug is not a high priority to straighten.

➡️ High Priority - Low Severity bug 
It could happen when, for example, the logo or name of the company is not displayed on the website. The bug is relevant to fix as soon as possible, although it may not cause enormous damage.

➡️ Low Priority - Low Severity bug 
For cases where the bug doesn’t cause a disaster and affects a low number of customers, both Severity and Priority are assigned low. For example, the privacy policy page takes a long time to load. A few people view the privacy policy page. Moreover, slow loading affects the customers much.